Report active tcp connections to configurable subnet.
When checking for active notebooks, if configured, will also check for active tcp connections to the configured subnet.
In the report, these IPs will be associated with the host computer. This is used as a proxy to see which computer is using which measurement device.

Data Protection Considerations:
This feature has abuse potential, if improperly configured and executed. To prevent this, only a small subnet (at most 2048 IP addresses) and only the bits covered by the hostmask are logged. This should make the data usable for the intended purpose (see above), but make it hard to use for nefarious purposes.